+++ util/convert_to_man.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+full_version=$(cargo run -- -V)
+version_number=$(echo "$full_version" | awk '{print $2}')
+
+touch spin.adoc
+
+asciidoctor -b manpage spin.adoc \
+  --destination=../man/ \
+  --attribute release-version="$version_number"