+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.2.0
 
 - Added a `--tail-command` flag to allow for continuous tailing the output of a specified command
+- Added a `-p` / `--print` flag to print the output directly to `stdout`
+- Added `man` pages
 
 ## 1.1.0 - 24.07.23